# api/openapi.yaml
openapi: 3.0.3
info:
  title: Varya One API
  version: 0.2.0
  description: Public REST contract for Varya One identity, company-scoped Cari, product catalog, pricing and tax foundations.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/setup:
    get:
      operationId: getSetupStatus
      summary: Read the permanent first-run setup state
      responses:
        "200":
          description: Setup state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupStatus" }
    post:
      operationId: completeSetup
      summary: Atomically create the first administrator, company, Merkez branch and Ana Depo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupRequest" }
      responses:
        "201":
          description: Setup completed and browser session created
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "409":
          {
            description: Setup is permanently closed,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Create a secure server-side browser session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          {
            description: Authenticated session,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/SessionResponse" } },
              },
          }
        "401":
          {
            description: Invalid credentials,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
        "429":
          {
            description: Login rate limited,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current browser session
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "204": { description: Session revoked }
        "401":
          {
            description: Authentication required,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
        "403":
          {
            description: CSRF validation failed,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
  /api/v1/session:
    get:
      operationId: getSession
      summary: Read the current user, companies and effective permissions
      responses:
        "200":
          {
            description: Current session,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/SessionResponse" } },
              },
          }
        "401":
          {
            description: Authentication required,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
  /api/v1/session/csrf:
    get:
      operationId: refreshSessionCSRF
      summary: Rotate the CSRF secret for the current browser session
      description: Requires an authenticated same-origin browser session. This endpoint changes only the session CSRF secret and does not change business data.
      security: [{ browserSession: [] }]
      responses:
        "200":
          description: New CSRF secret; the matching browser cookie is also set.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CSRFRefreshResponse" }
        "401":
          {
            description: Authentication required,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
        "403":
          {
            description: Origin or browser-session validation failed,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
  /api/v1/session/company:
    put:
      operationId: selectCompany
      summary: Select a company already authorized by active membership
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [company_id]
              properties: { company_id: { type: string, format: uuid } }
      responses:
        "200":
          {
            description: Updated session context,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/SessionResponse" } },
              },
          }
        "403":
          {
            description: Company access denied,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
...
  /api/v1/security/totp/setup:
    post:
      operationId: beginTOTPSetup
      summary: Create a pending encrypted TOTP secret
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "200":
          description: Secret displayed once for authenticator enrollment
          content:
            application/json:
              schema:
                type: object
                required: [secret, otpauth_uri]
                properties:
                  { secret: { type: string }, otpauth_uri: { type: string } }
  /api/v1/security/totp/confirm:
    post:
      operationId: confirmTOTPSetup
      summary: Verify pending TOTP and issue one-time recovery codes
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties: { code: { type: string, pattern: "^[0-9]{6}$" } }
      responses:
        "200":
          description: TOTP enabled and recovery codes displayed once
          content:
            application/json:
              schema:
                type: object
                required: [recovery_codes]
                properties:
                  { recovery_codes: { type: array, items: { type: string } } }
  /api/v1/api-tokens:
    get:
      operationId: listAPITokens
      summary: List metadata for API tokens in the selected company
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "200":
          {
            description: Token metadata; plaintext is never returned,
            content: { application/json: { schema: { type: object } } },
          }
    post:
      operationId: createAPIToken
      summary: Create a company-scoped API token displayed once
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/APITokenRequest" }
      responses:
        "201":
          {
            description: Token including one-time plaintext value,
            content: { application/json: { schema: { type: object } } },
          }
  /api/v1/api-tokens/{tokenID}:
    delete:
      operationId: revokeAPIToken
      summary: Revoke a token in the selected company
      parameters:
        - {
            name: tokenID,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "204": { description: Token revoked }
...
  /api/v1/roles:
    get:
      operationId: listRoles
      summary: List roles in the selected company
      responses:
        "200":
          {
            description: Company roles,
            content: { application/json: { schema: { type: object } } },
          }
    post:
      operationId: createRole
      summary: Create an editable company role
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleRequest" }
      responses:
        "201":
          {
            description: Role created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Role" } },
              },
          }
  /api/v1/roles/{roleID}:
    put:
      operationId: updateRole
      summary: Update a non-system role with optimistic locking
      parameters:
        - {
            name: roleID,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: If-Match,
            in: header,
            required: true,
            schema: { type: string },
          }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleRequest" }
      responses:
        "200":
          {
            description: Role updated,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Role" } },
              },
          }
        "412":
          {
            description: Version conflict,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
        "428":
          {
            description: If-Match required,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
...
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: varyaone_session
    apiToken:
      type: http
      scheme: bearer
      bearerFormat: Varya One scoped token
...
  schemas:
...
    SetupStatus:
      type: object
      additionalProperties: false
      required: [complete]
      properties: { complete: { type: boolean } }
    SetupRequest:
      type: object
      additionalProperties: false
      required:
        [admin_name, admin_email, password, legal_name, trade_name, entity_type]
      properties:
        admin_name: { type: string, minLength: 1, maxLength: 160 }
        admin_email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 1024 }
        legal_name: { type: string, minLength: 1, maxLength: 300 }
        trade_name: { type: string, minLength: 1, maxLength: 200 }
        entity_type: { type: string, enum: [LEGAL_ENTITY, SOLE_PROPRIETOR] }
        tax_number: { type: string, maxLength: 11 }
...
    SessionResponse:
      type: object
      required: [csrf_token, user, companies, expires_at, permissions]
      properties:
        csrf_token: { type: string }
        user:
          type: object
          required: [id, email, display_name, totp_enabled]
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            display_name: { type: string }
            totp_enabled: { type: boolean }
        companies:
          {
            type: array,
            items: { $ref: "#/components/schemas/CompanySummary" },
          }
        current_company_id: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
        permissions: { type: array, items: { type: string } }
...